Move record processing functions to separate module
